Fixed bug in prog. action bar title feature-kush
A bug in programmatically setting the title of action bar from user
login username was present - this meant that if nothing was passed or
when clicked on overview item in action bar, it showed 'null' or empty
string.  Fixed this bug - if nothing is passed, it will show just 'User
List' in OverView Activity instead of 'User List for null' or 'User List
for '.